add like query, made likeButton component
