+++ projects/bp-gallery/src/graphql/operation.graphql
@@ -172,6 +172,7 @@ query getPictureInfo($pictureId: ID!) {
             }
           }
         }
+        likes
       }
     }
   }
@@ -355,17 +356,6 @@ query getPicturesByAllSearch(
     }
   }
 }
-
-#query getPictureLikes($pictureId: ID!) {
-#  picture(id: $pictureId) {
-#    data {
-#      id
-#      attributes {
-#        likes
-#      }
-#    }
-#  }
-#}
 
 query getCollectionInfoByName($collectionName: String, $publicationState: PublicationState = LIVE) {
   collections(filters: { name: { eq: $collectionName } }, publicationState: $publicationState) {
